requirements.txt
chromadb==0.4.24
fastapi==0.110.1
langchain_community==0.0.31
langchain_core==0.1.39
langchain-experimental==0.0.56
langchain_openai==0.1.1
pydantic==2.6.4
pypdf==4.1.0
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.32.2
